+++ workflows_definition/src/openapi/closing_reason_sdk.py
@@ -89,8 +89,8 @@ class ClosingReasonSDK(BaseSDK):
         if utils.match_response(http_res, "202", "*"):
             return
         if utils.match_response(http_res, ["400", "500"], "application/json"):
-            data = utils.unmarshal_json(http_res.text, models.ErrorRespData)
-            raise models.ErrorResp(data=data)
+            data = utils.unmarshal_json(http_res.text, models.ErrorRespErrorData)
+            raise models.ErrorRespError(data=data)
         if utils.match_response(http_res, ["4XX", "5XX"], "*"):
             raise models.SDKError(
                 "API error occurred", http_res.status_code, http_res.text, http_res
@@ -185,8 +185,8 @@ class ClosingReasonSDK(BaseSDK):
         if utils.match_response(http_res, "202", "*"):
             return
         if utils.match_response(http_res, ["400", "500"], "application/json"):
-            data = utils.unmarshal_json(http_res.text, models.ErrorRespData)
-            raise models.ErrorResp(data=data)
+            data = utils.unmarshal_json(http_res.text, models.ErrorRespErrorData)
+            raise models.ErrorRespError(data=data)
         if utils.match_response(http_res, ["4XX", "5XX"], "*"):
             raise models.SDKError(
                 "API error occurred", http_res.status_code, http_res.text, http_res